Fix APIGatewayProxyEvent headers fields
When no headers are present on the incoming HTTP request, API Gateway sends `null` to the lambda under `headers` and `multiValueHeaders`. This change reflects that possible null value.

Question about procedure: this is likely a breaking change for consumers. Does that imply a major version bump for the npm package?